+++ core-ng/src/main/java/core/framework/http/HTTPClientBuilder.java
@@ -1,7 +1,7 @@
 package core.framework.http;
 
-import core.framework.impl.http.HTTPClientImpl;
-import core.framework.impl.http.TrustAllTrustManager;
+import core.framework.internal.http.DefaultTrustManager;
+import core.framework.internal.http.HTTPClientImpl;
 import core.framework.util.StopWatch;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,10 +20,25 @@ import java.time.Duration;
  * @author neo
  */
 public final class HTTPClientBuilder {
+    static {
+        // the system property must be set before loading jdk http client class, JDK uses static field to initialize those values
+
+        // allow server ssl cert change during renegotiation
+        // http client uses pooled connection, multiple requests to same host may hit different server behind LB
+        // refer to sun.security.ssl.ClientHandshakeContext, allowUnsafeServerCertChange = Utilities.getBooleanProperty("jdk.tls.allowUnsafeServerCertChange", false);
+        System.setProperty("jdk.tls.allowUnsafeServerCertChange", "true");
+
+        // api client keep alive should be shorter than server side in case server side disconnect connection first
+        // refer to jdk.internal.net.http.ConnectionPool,
+        // jdk.internal.net.http.HttpClientImpl.SelectorManager.DEF_NODEADLINE uses 3s as interval to check
+        System.setProperty("jdk.httpclient.keepalive.timeout", "30");   // 30s timeout for keep alive
+    }
+
     private final Logger logger = LoggerFactory.getLogger(HTTPClientBuilder.class);
 
+    private Duration connectTimeout = Duration.ofSeconds(25);
     private Duration timeout = Duration.ofSeconds(60);
-    private Duration slowOperationThreshold = Duration.ofSeconds(30);
+    private Duration slowOperationThreshold = Duration.ofSeconds(30);   // slow threshold should be greater than connect timeout, to let connect timeout happen first
     private boolean enableCookie = false;
     private boolean enableRedirect = false;
     private int maxRetries = 1;
@@ -37,21 +52,25 @@ public final class HTTPClientBuilder {
 
             if (trustAll) {
                 SSLContext sslContext = SSLContext.getInstance("TLS");
-                sslContext.init(null, new TrustManager[]{new TrustAllTrustManager()}, new SecureRandom());
+                sslContext.init(null, new TrustManager[]{new DefaultTrustManager()}, new SecureRandom());
                 builder.sslContext(sslContext);
             }
 
-            builder.connectTimeout(timeout);
+            builder.connectTimeout(connectTimeout);
             if (enableRedirect) builder.followRedirects(HttpClient.Redirect.NORMAL);
             if (enableCookie) builder.cookieHandler(new CookieManager(null, CookiePolicy.ACCEPT_ALL));
-            HttpClient httpClient = builder.build();
 
-            return new HTTPClientImpl(httpClient, userAgent, timeout, maxRetries, slowOperationThreshold);
+            return new HTTPClientImpl(builder.build(), userAgent, timeout, maxRetries, slowOperationThreshold);
         } catch (NoSuchAlgorithmException | KeyManagementException e) {
             throw new Error(e);
         } finally {
             logger.info("create http client, elapsed={}", watch.elapsed());
         }
+    }
+
+    public HTTPClientBuilder connectTimeout(Duration connectTimeout) {
+        this.connectTimeout = connectTimeout;
+        return this;
     }
 
     public HTTPClientBuilder timeout(Duration timeout) {