+++ core-ng/src/main/java/core/framework/impl/web/ShutdownHandler.java
@@ -13,9 +13,9 @@ import java.util.concurrent.atomic.AtomicLong;
  * @author neo
  */
 public class ShutdownHandler implements ExchangeCompletionListener {
+    public final AtomicBoolean shutdown = new AtomicBoolean(false);
     final AtomicLong activeRequests = new AtomicLong(0);
     private final Logger logger = LoggerFactory.getLogger(ShutdownHandler.class);
-    private final AtomicBoolean shutdown = new AtomicBoolean(false);
     private final Object lock = new Object();
 
     boolean handle(Exchange exchange) {
@@ -24,16 +24,13 @@ public class ShutdownHandler implements ExchangeCompletionListener {
 
         if (shutdown.get()) {
             logger.warn("reject request due to server is shutting down, requestURL={}", exchange.getRequestURL());
+            exchange.setPersistent(false);  // ask client not set keep alive for current connection, with persistent=false, undertow will send header "connection: close"
             exchange.setStatusCode(StatusCodes.SERVICE_UNAVAILABLE);
             exchange.endExchange();
             return true;
         }
 
         return false;
-    }
-
-    void shutdown() {
-        shutdown.set(true);
     }
 
     boolean awaitTermination(long timeoutInMs) throws InterruptedException {