+++ core-ng/src/main/java/core/framework/impl/web/service/WebServiceClient.java
@@ -95,14 +95,12 @@ public class WebServiceClient {
         if (status.code >= 200 && status.code < 300) return;
         try {
             ErrorResponse error = (ErrorResponse) responseBeanMapper.fromJSON(ErrorResponse.class, response.body);
-            logger.debug("failed to call remote service, id={}, severity={}, errorCode={}, remoteStackTrace={}", error.id, error.severity, error.errorCode, error.stackTrace);
-            throw new RemoteServiceException(error.message, parseSeverity(error.severity), error.errorCode, status);
+            logger.debug("failed to call remote service, status={}, id={}, severity={}, errorCode={}, remoteStackTrace={}", status, error.id, error.severity, error.errorCode, error.stackTrace);
+            throw new RemoteServiceException(format("failed to call remote service, status={}, error={}", status, error.message), parseSeverity(error.severity), error.errorCode, status);
         } catch (RemoteServiceException e) {
             throw e;
         } catch (Throwable e) {
-            String responseText = response.text();
-            logger.warn("failed to decode response, statusCode={}, responseText={}", status.code, responseText, e);
-            throw new RemoteServiceException(format("internal communication failed, status={}, responseText={}", status.code, responseText), Severity.ERROR, "REMOTE_SERVICE_ERROR", status, e);
+            throw new RemoteServiceException(format("internal communication failed, status={}, responseText={}", status.code, response.text()), Severity.ERROR, "REMOTE_SERVICE_ERROR", status, e);
         }
     }