+++ core-ng/src/main/java/core/framework/internal/stat/Metrics.java
@@ -1,0 +1,10 @@
+package core.framework.internal.stat;
+
+import java.util.Map;
+
+/**
+ * @author neo
+ */
+public interface Metrics {
+    void collect(Map<String, Double> stats);
+}