+++ core-ng/src/test/java/core/framework/impl/web/ShutdownHandlerTest.java
@@ -31,10 +31,11 @@ class ShutdownHandlerTest {
 
     @Test
     void handleShutdown() {
-        handler.shutdown();
+        handler.shutdown.set(true);
         assertThat(handler.handle(exchange)).isTrue();
 
         verify(exchange).setStatusCode(StatusCodes.SERVICE_UNAVAILABLE);
+        verify(exchange).setPersistent(false);
         verify(exchange).endExchange();
     }